Make Soong integration tests use hermetic toybox tools
Use the hermetic toybox tools when running Soong integration tests
using run_ci_tests.sh.  This should fix some of the problems the
mac versions of the tests are hitting.

Bug: 450580777
Test: run_ci_tests.sh

+++ tests/lib.sh
@@ -11,11 +11,11 @@ HARDWIRED_MOCK_TOP=
 REAL_TOP="$(readlink -f "$(dirname "$0")"/../../..)"
 
 function make_mock_top {
-  mock=$(mktemp -t -d st.XXXXX)
+  mock=$(mktemp -t -d st.XXXXXX)
   echo "$mock"
 }
 
-WARMED_UP_MOCK_TOP=$(mktemp -t soong_integration_tests_warmup.XXXXXX.tar.gz)
+WARMED_UP_MOCK_TOP=$(mktemp -t soong_integration_tests_warmup.tar.gz.XXXXXX)
 MOCK_TOP_TO_CLEAN_UP=
 if [[ -n "$HARDWIRED_MOCK_TOP" ]]; then
   MOCK_TOP="$HARDWIRED_MOCK_TOP"
@@ -50,7 +50,7 @@ function warmup_mock_top {
 
 function cleanup_mock_top {
   cd /
-  if [[ -d "$MOCK_TOP_TO_CLEAN_UP" ]]; then
+  if [[ -n "$MOCK_TOP_TO_CLEAN_UP" && -d "$MOCK_TOP_TO_CLEAN_UP" ]]; then
     rm -fr "$MOCK_TOP_TO_CLEAN_UP"
   fi
 }
